♻️ refactor TransactionService + added TransactionRepository
