+++ src/populate_neo.py
@@ -4,7 +4,7 @@ import os
 from src.services.TransactionService import TransactionService
 from src.models.neo4j_models import Transaction
 from src.logs.logger import DataProcessingLogger
-from dotenv import load_dotenv
+from src.database.neo4j import Neo4jDatabase
 
 
 #SCRIPT TO INSERT FIRST DATA FROM LOG FILE TO NEO4J
@@ -15,12 +15,9 @@ def main():
         for line in file:
             data.append(json.loads(line))
 
-    load_dotenv()
-    neo4j_uri = os.getenv("NEO4J_URI")
-    neo4j_user = os.getenv("NEO4J_USER")
-    neo4j_password = os.getenv("NEO4J_PASSWORD")
 
-    transaction_service = TransactionService(neo4j_uri, neo4j_user, neo4j_password)
+    driver = Neo4jDatabase.get_instance().driver
+    transaction_service = TransactionService(driver)
 
 
     total_items = len(data)
@@ -32,7 +29,6 @@ def main():
             transaction_id=item['id'],
             nft=item['nft']['id']
         )
-        transaction_service.processTransaction(transaction)
         transaction_service.processTransaction(transaction)
